# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/address": {
            "get": {
                "tags": [
                    "address"
                ],
                "summary": "Returns a list of adress(es) related to an user",
                "description": "Returns a map of status codes to quantities",
                "operationId": "getAddresses",
                "parameters": [
                    {
                        "name": "userHashID",
                        "in": "path",
                        "description": "ID of the users that owns the adress(es)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "operação realizado com sucesso"
                    }
                }
            }
        },
        "/address/": {
            "post": {
                "tags": [
                    "address"
                ],
                "summary": "Add a new address.",
                "operationId": "createAddress",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "rua",
                                    "cidade",
                                    "estado",
                                    "cep",
                                    "user_id"
                                ],
                                "properties": {
                                    "rua": {
                                        "type": "text"
                                    },
                                    "cidade": {
                                        "type": "text"
                                    },
                                    "estado": {
                                        "type": "password"
                                    },
                                    "cep": {
                                        "type": "text"
                                    },
                                    "user_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Erro ao criar novo endereço"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Returns user details and authentication",
                "description": "Returns a map of status codes to quantities",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "give the user logout acess.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Returns a list of user(es)",
                "description": "Returns a map of status codes to quantities",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "operação realizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Returns a user",
                "description": "Returns a map of status codes to quantities",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "userHashID",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "usuario encontrado com sucesso"
                    },
                    "422": {
                        "description": "Desculpe, Nao foi possivel encontrar usuario"
                    }
                }
            }
        },
        "/users/": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "create a user",
                "description": "Returns a map of status codes to quantities",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "full_name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "full_name": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "usuario criado com sucesso"
                    },
                    "422": {
                        "description": "Erro ao criar novo usuario"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "users/{id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "upadate a user",
                "description": "Returns a map of status codes to quantities",
                "operationId": "update",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "full_name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "full_name": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario criado com sucesso"
                    },
                    "422": {
                        "description": "Erro ao criar novo usuario"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "delete a user",
                "description": "Returns a map of status codes to quantities",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "userHashId",
                        "in": "path",
                        "description": "user hash id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "usuario deletado com sucesso"
                    },
                    "422": {
                        "description": "Desculpe, o usuario não pode ser excluido."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}